Move space page related tests into a space-page cypress folder
