+++ ui/cypress/e2e/space-page/form_field_dropdowns.cy.ts
